Update WP 5.0.3 and Plugins

+++ wp-content/plugins/search-by-algolia-instant-relevant-results/includes/indices/class-algolia-searchable-posts-index.php
@@ -79,7 +79,7 @@ final class Algolia_Searchable_Posts_Index extends Algolia_Index {
 
 		$removed = remove_filter( 'the_content', 'wptexturize', 10 );
 
-		$post_content = apply_filters( 'algolia_searchable_post_content', $post->post_content );
+		$post_content = apply_filters( 'algolia_searchable_post_content', $post->post_content, $post );
 		$post_content = apply_filters( 'the_content', $post_content );
 
 		if ( true === $removed ) {
@@ -243,10 +243,17 @@ final class Algolia_Searchable_Posts_Index extends Algolia_Index {
 	 * @param array   $records
 	 */
 	private function update_post_records( WP_Post $post, array $records ) {
-		$this->delete_item( $post );
+		// If there are no records, parent `update_records` will take care of the deletion.
+		// In case of posts, we ALWAYS need to delete existing records.
+		if ( ! empty( $records ) ) {
+			$this->delete_item( $post );
+		}
 
-		// Update the other records.
 		parent::update_records( $post, $records );
+
+		// Keep track of the new record count for future updates relying on the objectID's naming convention .
+		$new_records_count = count( $records );
+		$this->set_post_records_count( $post, $new_records_count );
 
 		do_action( 'algolia_searchable_posts_index_post_updated', $post, $records );
 		do_action( 'algolia_searchable_posts_index_post_' . $post->post_type . '_updated', $post, $records );
@@ -307,10 +314,31 @@ final class Algolia_Searchable_Posts_Index extends Algolia_Index {
 	 */
 	public function delete_item( $item ) {
 		$this->assert_is_supported( $item );
-		$this->get_index()->deleteBy(
-			array(
-				'filters' => 'post_id=' . $item->ID,
-			)
-		);
+
+		$records_count = $this->get_post_records_count( $item->ID );
+		$object_ids    = array();
+		for ( $i = 0; $i < $records_count; $i++ ) {
+			$object_ids[] = $this->get_post_object_id( $item->ID, $i );
+		}
+
+		if ( ! empty( $object_ids ) ) {
+			$this->get_index()->deleteObjects( $object_ids );
+		}
+	}
+
+	/**
+	 * @param int $post_id
+	 *
+	 * @return int
+	 */
+	private function get_post_records_count( $post_id ) {
+		return (int) get_post_meta( (int) $post_id, 'algolia_' . $this->get_id() . '_records_count', true );
+	}
+	/**
+	 * @param WP_Post $post
+	 * @param int     $count
+	 */
+	private function set_post_records_count( WP_Post $post, $count ) {
+		update_post_meta( (int) $post->ID, 'algolia_' . $this->get_id() . '_records_count', (int) $count );
 	}
 }
